+++ src/Plugin.php
@@ -12,12 +12,14 @@ use craft\base\Plugin as BasePlugin;
 use craft\events\RegisterUrlRulesEvent;
 use craft\helpers\UrlHelper;
 use craft\log\FileTarget;
+use craft\web\twig\variables\CraftVariable;
 use craft\web\UrlManager;
 use venveo\oauthclient\services\Apps as AppsService;
 use venveo\oauthclient\services\Credentials as CredentialsService;
 use venveo\oauthclient\services\Providers;
 use venveo\oauthclient\services\Providers as ProvidersService;
 use venveo\oauthclient\services\Tokens as TokensService;
+use venveo\oauthclient\variables\OAuthVariable;
 use yii\base\Event;
 
 /**
@@ -67,6 +69,7 @@ class Plugin extends BasePlugin
         $this->_registerLogger();
         $this->_setComponents();
         $this->_registerCpRoutes();
+        $this->_registerVariables();
     }
 
     // Protected Methods
@@ -123,5 +126,18 @@ class Plugin extends BasePlugin
                 'oauthclient/authorize/<handle:{handle}>' => 'oauthclient/authorize/authorize-app',
             ]);
         });
+    }
+
+    private function _registerVariables()
+    {
+        Event::on(
+            CraftVariable::class,
+            CraftVariable::EVENT_INIT,
+            function (Event $event) {
+                /** @var CraftVariable $variable */
+                $variable = $event->sender;
+                $variable->set('oauth', OAuthVariable::class);
+            }
+        );
     }
 }